Move mod_integrity to separate module
