+++ src/framework/mod/core.c
@@ -13,7 +13,6 @@
 #include <assert.h>
 
 #include <util/array.h>
-#include <lib/crypt/md5.h>
 #include <framework/mod/api.h>
 #include <framework/mod/ops.h>
 #include <framework/mod/types.h>
@@ -94,27 +93,6 @@ int mod_activate_app(const struct mod *mod) {
 		memcpy(app->data, app->data + APP_DATA_RESERVE_OFFSET, app->data_sz);
 		memset(app->bss, 0, app->bss_sz);
 		*mod->priv = priv;
-	}
-
-	return 0;
-}
-
-int mod_integrity_check(const struct mod *mod) {
-	const struct mod_label *label = mod_label(mod);
-	unsigned char md5[16];
-
-	if (!label || !label->text.md5sum) {
-		return -ENOTSUP;
-	}
-
-	md5_count((unsigned char *) label->text.vma, label->text.len, md5);
-	if (0 != memcmp(md5, label->text.md5sum, 16)) {
-		return 1;
-	}
-
-	md5_count((unsigned char *) label->rodata.vma, label->rodata.len, md5);
-	if (0 != memcmp(md5, label->rodata.md5sum, 16)) {
-		return 1;
 	}
 
 	return 0;
